\name{ssm_sim}
\alias{ssm_sim}
\title{Simulate states from a model}
\arguments{
  \item{object}{An object containing MCMC samples from a
  state space model.}

  \item{data}{\code{list} Additional data needed by the
  method, usually the observations.}
}
\value{
  An object of class \code{"array"}.
}
\description{
  Simulate states from a model
}

