\name{ssm_sim}
\alias{ssm_sim}
\title{Simulate states}
\arguments{
  \item{object}{An object containing MCMC samples from an
  object of class \code{SdDLM}}

  \item{data}{\code{list} Additional data needed by the
  method, usually the observations.}
}
\value{
  An object of class \code{"array"}.
}
\description{
  Simulate states
}

